Videos, links, rotations
Figured out the codec or something was the reason the flipped videos didn't load in
android. Using h264/aac instead of mpeg4. Uploaded super low quality videos to dropbox
and updated links accordingly in attempt to avoid buffering issues (1 MB!!! SO LOW).
Rotation values seem to be failing. Updating according, but in progress.